Move enum to the right place in the structure.
