+++ src/main/java/com/gra/local/persistence/domain/VendorAccount.java
@@ -1,6 +1,5 @@
 package com.gra.local.persistence.domain;
 
-import com.gra.local.persistence.services.dtos.Role;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
